testing: fix all test specs that broke after adding OS boot.
when we added the OS boot, it immediately jumped to the program offset in the same cycle. however, tests were still invoking System#cycle() again, which lead to continue executing all bytes until it reached past memory limits. this went undetected since we were not surfacing errors in a meaningful way.

includes:
- test fixes
- strict checks for invalid memory access
- surfacing critical errors in the js console